+++ packages/desktop/src/rtl-tests/possibleImpersonationAttack.test.tsx
@@ -99,7 +99,9 @@ describe('Possible Impersonation Attack', () => {
     const modal = screen.getByTestId('possible-impersonation-attack-modal-component')
     expect(modal).toBeVisible()
 
-    const button = screen.getByTestId('unregistered-button')
-    expect(button).toBeVisible()
+    // Temporarily hiding button - https://github.com/TryQuiet/quiet/issues/2025 */
+
+    // const button = screen.getByTestId('unregistered-button')
+    // expect(button).toBeVisible()
   })
 })